TodoList.web
"I am happy to be able to complete the task of making this todolist well. This project provides a very valuable experience to improve my programming capabilities, especially in the use of JS programming languages. I also learn a lot about the basic concepts of making applications, such as models-view-controller ( MVC) and effective debugging techniques.